ft(user) -implement user login and signup -[WIP ft-user]
